Add CSS minification + error handling
Error handling is kinda terribly implemented right now... I originally
let the underlying JS throw an error and then strip the message.
However, in some cases it seemed the V8 engine printed the original
error to the console anyway, so I opted for the messy variant in this
commit. Would ideally like to revisit.