Bugfix: Regexp fail in cache-busting replacement
Regexp replacement certainly leads to many edge cases. Usually the file
extension makes for a natural ending delimiter, but that goes out the
window with the 'woff' and 'woff2' file formats.

Say that you have /fonts/awesome.woff and /fonts/awesome.woff2 - the
former is entirely contained in the latter. This makes for some tricky
replacements. My solution at this time is to always replace the longer
strings first.

Closes #43