Make asset loading work on Windows
The class path parser is changed so that it can parse out the class
path on Windows, which looks different than the class path on *nix. The
JVM provides a handy (File/pathSeparator) we can use, so no OS detection
is needed.

The asset creation is tweaked to remove the assumption that file system
paths look like web URLs. A separate step is added, before we run the
emacs file artefact filter and match the paths against regexps, which
converts the file system paths to URL-like paths. No OS detection is
needed here either, as we rely on java.nio.Path which has a handy
method to list out all the segments in a path, cross platform.