Add support for globs as well as regexps for path matching
We piggyback on the built-in globbing support in java.nio.

A flexible format is added to load-assets and load-bundle, which takes
the form of `[:format-type & args]`. A plain string is converted into
`[:path str]` and a regex is converted to `[:regex r]`. Additionally,
you can pass in `[:glob glob-pattern]`, which is how you use the new
globbing functionality.

The full globbing syntax is documented here:

https://docs.oracle.com/javase/7/docs/api/java/nio/file/FileSystem.html#getPathMatcher(java.lang.String)